Fixed deletes of all entities related to the new groups, also implemented gc mechanism for new groups

+++ src/main/java/net/sourceforge/fenixedu/domain/organizationalStructure/PersonFunction.java
@@ -223,6 +223,9 @@ public class PersonFunction extends PersonFunction_Base {
         if (hasSender()) {
             setSender(null);
         }
+        if (!getDelegateStudentsGroupSet().isEmpty()) {
+            throw new DomainException("error.personFunction.cannotDeletePersonFunctionUsedInAccessControl");
+        }
         setExecutionInterval(null);
         super.delete();
     }
